(raptor_xml_writer_indent): Cast for signed/unsigned int compare.

+++ src/raptor_xml_writer.c
@@ -139,7 +139,7 @@ raptor_xml_writer_indent(raptor_xml_writer *xml_writer)
   
   while (num_spaces > 0) {
 
-    int count = (num_spaces > SPACES_BUFFER_SIZE) ? SPACES_BUFFER_SIZE : num_spaces;
+    int count = (num_spaces > SPACES_BUFFER_SIZE) ? (int)SPACES_BUFFER_SIZE : num_spaces;
 
     raptor_iostream_write_counted_string(xml_writer->iostr, spaces_buffer, count);
